+++ lib/cldr/number/format/options.ex
@@ -17,7 +17,6 @@ defmodule Cldr.Number.Format.Options do
     :locale,
     :number_system,
     :currency,
-    :scientific,
     :format,
     :currency_format,
     :currency_digits,
@@ -397,10 +396,6 @@ defmodule Cldr.Number.Format.Options do
           {:error, _} -> {:error, Cldr.unknown_currency_error(currency)}
         end
     end
-  end
-
-  defp validate_option(:scientific, _options, _backend, nil) do
-    {:ok, :scientific}
   end
 
   # If a currency code is provided but no format then a currency