+++ lib/cldr/number/formatter/decimal_formatter.ex
@@ -826,17 +826,17 @@ defmodule Cldr.Number.Formatter.Decimal do
 
   def adjust_fraction_for_currency(meta, currency, :accounting, backend) do
     {:ok, currency} = Currency.currency_for_code(currency, backend)
-    do_adjust_fraction(meta, currency.digits, currency.rounding)
+    do_adjust_fraction(meta, meta.fractional_digits.max, currency.rounding)
   end
 
   def adjust_fraction_for_currency(meta, currency, :cash, backend) do
     {:ok, currency} = Currency.currency_for_code(currency, backend)
-    do_adjust_fraction(meta, currency.cash_digits, currency.cash_rounding)
+    do_adjust_fraction(meta, meta.fractional_digits.max, currency.cash_rounding)
   end
 
   def adjust_fraction_for_currency(meta, currency, :iso, backend) do
     {:ok, currency} = Currency.currency_for_code(currency, backend)
-    do_adjust_fraction(meta, currency.iso_digits, currency.iso_digits)
+    do_adjust_fraction(meta, meta.fractional_digits.max, currency.iso_digits)
   end
 
   defp do_adjust_fraction(meta, digits, rounding) do